# docs/myrctf-cockpit/openapi.yaml
openapi: 3.1.0
info:
  title: CTFd Cockpit Ingest API
  version: 1.0.0
  description: Sensitive, gzip-only replication API. Request bodies and credentials must never be logged.
servers:
  - url: https://myrctf-telemetry.3rr.dev
security:
  - bearerAuth: []
paths:
  /api/v1/events:
    post:
      operationId: ingestEvents
      parameters: [{$ref: '#/components/parameters/Gzip'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: schemas/v1/event-batch.schema.json}
      responses:
        '200': {$ref: '#/components/responses/EventCommitted'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '413': {$ref: '#/components/responses/Error'}
        '415': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
  /api/v1/snapshot:
    post:
      operationId: ingestSnapshotChunk
      parameters: [{$ref: '#/components/parameters/Gzip'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: schemas/v1/snapshot-batch.schema.json}
      responses:
        '200':
          description: Chunk committed; reconciliation occurs only after a contiguous complete chunk.
          content:
            application/json:
              schema: {type: object, additionalProperties: false}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '413': {$ref: '#/components/responses/Error'}
        '415': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
  /healthz:
    get:
      security: []
      responses: {'200': {description: Process alive}}
  /readyz:
    get:
      security: []
      responses:
        '200': {description: Database and schema ready}
        '503': {description: Not ready}
  /metrics:
    get:
      security: []
      responses:
        '200':
          description: Prometheus exposition
          content: {text/plain: {schema: {type: string}}}
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer}
  parameters:
    Gzip:
      name: Content-Encoding
      in: header
      required: true
      schema: {type: string, const: gzip}
  responses:
    EventCommitted:
      description: Transaction committed. The checkpoint is exactly the received integer.
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [checkpoint]
            properties: {checkpoint: {type: integer, minimum: 1}}
    Error:
      description: Request rejected without partial writes.
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [error]
            properties: {error: {type: string}}